fix events api and test

+++ backend-node/src/controllers/eventController.js
@@ -1,4 +1,5 @@
 const pool = require("../config/db");
+const calendarController = require("./calendarController");
 
 exports.getEventsByCalendar = async (req, res) => {
   const { calendarId } = req.params;
@@ -6,17 +7,7 @@ exports.getEventsByCalendar = async (req, res) => {
 
   try {
     // Permission check
-    const hasPermission = await pool.query(
-      `SELECT 1 FROM calendars 
-       WHERE id = $1 
-       AND (owner_id = $2 OR EXISTS (
-         SELECT 1 FROM calendar_shared_users 
-         WHERE calendar_id = $1 AND user_id = $2
-       ))`,
-      [calendarId, userId]
-    );
-
-    if (hasPermission.rowCount === 0) {
+    if (!(await calendarController.calendarPermission(calendarId, userId))) {
       return res.status(403).json({ error: "Permission denied" });
     }
 
@@ -24,9 +15,9 @@ exports.getEventsByCalendar = async (req, res) => {
       "SELECT id, title, start_time, end_time FROM events WHERE calendar_id = $1 ORDER BY start_time",
       [calendarId]
     );
-    res.status(201).json(result.rows);
+    res.status(200).json(result.rows);
   } catch (err) {
-    res.status(500).json({ error: "Get fails", details: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
 
@@ -39,13 +30,16 @@ exports.createEvent = async (req, res) => {
   }
 
   try {
+    if (!(await calendarController.calendarPermission(calendar_id, userId))) {
+      return res.status(403).json({ error: "Permission denied" });
+    }
     const result = await pool.query(
       "INSERT INTO events (calendar_id, title, start_time, end_time) VALUES ($1, $2, $3, $4) RETURNING id, title, start_time, end_time",
       [calendar_id, title, start_time, end_time]
     );
     res.status(201).json(result.rows[0]);
   } catch (err) {
-    res.status(500).json({ error: "Create fails", details: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
 
@@ -53,45 +47,53 @@ exports.updateEvent = async (req, res) => {
   const { id } = req.params;
   const { title, start_time, end_time } = req.body;
   const userId = req.user.id;
+
   try {
+    const calendarId = await findCalendarIdbyEventId(id);
     // Permission check
+    if (!(await calendarController.calendarPermission(calendarId, userId))) {
+      return res.status(403).json({ error: "Permission denied" });
+    }
     const result = await pool.query(
       `
       UPDATE events 
       SET title = $1, start_time = $2, end_time = $3, updated_at = NOW()
-      FROM calendars
-      WHERE events.id = $4 
-        AND events.calendar_id = calendars.id
-        AND (calendars.owner_id = $5 
-             OR EXISTS (
-               SELECT 1 FROM calendar_shared_users 
-               WHERE calendar_id = calendars.id 
-               AND user_id = $5 
-               AND permission IN ('write')
-             )
-      RETURNING events.*
+      WHERE id = $4 
+      RETURNING *
     `,
-      [title, start_time, end_time, id, userId]
+      [title, start_time, end_time, id]
     );
 
     if (result.rowCount === 0) {
-      return res
-        .status(404)
-        .json({ error: "Event not existing or permission denied" });
+      return res.status(404).json({ error: "Event not existing" });
     }
 
-    res.status(201).json(result.rows[0]);
+    res.status(200).json(result.rows[0]);
   } catch (err) {
-    res.status(500).json({ error: "Update fails", details: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
 
 exports.deleteEvent = async (req, res) => {
   const { id } = req.params;
+  const userId = req.user.id;
   try {
+    const calendarId = await findCalendarIdbyEventId(id);
+    // Permission check
+    if (!(await calendarController.calendarPermission(calendarId, userId))) {
+      return res.status(403).json({ error: "Permission denied" });
+    }
     await pool.query("DELETE FROM events WHERE id = $1", [id]);
     res.status(204).end();
   } catch (err) {
-    res.status(500).json({ error: "Delete fails", details: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
+
+async function findCalendarIdbyEventId(eventId) {
+  const result = await pool.query(
+    "SELECT calendar_id FROM events WHERE id = $1",
+    [eventId]
+  );
+  return result.rows[0].calendar_id;
+}
